Refactor: move `MockClientCertVerifier` to common test module
