+++ rustls/tests/client_cert_verifier.rs
@@ -8,18 +8,13 @@ mod common;
 use std::sync::Arc;
 
 use common::{
-    do_handshake_until_both_error, do_handshake_until_error, get_client_root_store,
-    make_client_config_with_versions, make_client_config_with_versions_with_auth,
-    make_pair_for_arc_configs, server_config_builder, server_name, webpki_client_verifier_builder,
-    ErrorFromPeer, KeyType, ALL_KEY_TYPES,
+    do_handshake_until_both_error, do_handshake_until_error, make_client_config_with_versions,
+    make_client_config_with_versions_with_auth, make_pair_for_arc_configs, server_config_builder,
+    server_name, ErrorFromPeer, KeyType, MockClientVerifier, ALL_KEY_TYPES,
 };
-use pki_types::{CertificateDer, UnixTime};
-use rustls::client::danger::HandshakeSignatureValid;
-use rustls::internal::msgs::handshake::DistinguishedName;
-use rustls::server::danger::{ClientCertVerified, ClientCertVerifier};
+use rustls::server::danger::ClientCertVerified;
 use rustls::{
-    AlertDescription, ClientConnection, DigitallySignedStruct, Error, InvalidMessage, ServerConfig,
-    ServerConnection, SignatureScheme,
+    AlertDescription, ClientConnection, Error, InvalidMessage, ServerConfig, ServerConnection,
 };
 
 // Client is authorized!
@@ -134,76 +129,6 @@ fn client_verifier_fails_properly() {
                 err,
                 Err(ErrorFromPeer::Server(Error::General("test err".into())))
             );
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct MockClientVerifier {
-    parent: Arc<dyn ClientCertVerifier>,
-    pub verified: fn() -> Result<ClientCertVerified, Error>,
-    pub subjects: Vec<DistinguishedName>,
-    pub mandatory: bool,
-    pub offered_schemes: Option<Vec<SignatureScheme>>,
-}
-
-impl MockClientVerifier {
-    pub fn new(verified: fn() -> Result<ClientCertVerified, Error>, kt: KeyType) -> Self {
-        Self {
-            parent: webpki_client_verifier_builder(get_client_root_store(kt))
-                .build()
-                .unwrap(),
-            verified,
-            subjects: get_client_root_store(kt).subjects(),
-            mandatory: true,
-            offered_schemes: None,
-        }
-    }
-}
-
-impl ClientCertVerifier for MockClientVerifier {
-    fn client_auth_mandatory(&self) -> bool {
-        self.mandatory
-    }
-
-    fn root_hint_subjects(&self) -> &[DistinguishedName] {
-        &self.subjects
-    }
-
-    fn verify_client_cert(
-        &self,
-        _end_entity: &CertificateDer<'_>,
-        _intermediates: &[CertificateDer<'_>],
-        _now: UnixTime,
-    ) -> Result<ClientCertVerified, Error> {
-        (self.verified)()
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        message: &[u8],
-        cert: &CertificateDer<'_>,
-        dss: &DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, Error> {
-        self.parent
-            .verify_tls12_signature(message, cert, dss)
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        message: &[u8],
-        cert: &CertificateDer<'_>,
-        dss: &DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, Error> {
-        self.parent
-            .verify_tls13_signature(message, cert, dss)
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<SignatureScheme> {
-        if let Some(schemes) = &self.offered_schemes {
-            schemes.clone()
-        } else {
-            self.parent.supported_verify_schemes()
         }
     }
 }
